+++ crawler/src/RequestManager.cpp
@@ -90,6 +90,9 @@ void RequestManager::Run(ThreadSync& sync) {
         // Process ready responses
         auto& ready = requestExecutor_.ReadyResponses();
         if (!ready.empty()) {
+            for (const auto& r : ready) {
+                limiter_->UnleaseHost(r.req.Url().host, r.req.Url().NonEmptyPort());
+            }
             docQueue_->PushAll(ready);
             ready.clear();
         }
@@ -112,6 +115,14 @@ void RequestManager::TouchRequestTimeouts() {
 }
 
 void RequestManager::DispatchFailedRequest(const http::FailedRequest& failed) {
+<<<<<<< HEAD
+||||||| ac22a20
+    limiter_->UnleaseHost(failed.req.Url().host);
+
+=======
+    limiter_->UnleaseHost(failed.req.Url().host, failed.req.Url().NonEmptyPort());
+
+>>>>>>> origin/main
     auto s = std::string{http::StringOfRequestError(failed.error)};
     spdlog::warn("failed crawl request: {} {}", failed.req.Url().url, s);